add second client, initial tests

+++ src/main/java/com/booksmoviesapp/domain/dto/ReviewDto.java
@@ -2,13 +2,7 @@ package com.booksmoviesapp.domain.dto;
 
 import com.booksmoviesapp.domain.Book;
 import com.booksmoviesapp.domain.Movie;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-//@Getter
-//@AllArgsConstructor
-//@NoArgsConstructor
 public class ReviewDto {
 
     private Long reviewId;
@@ -73,7 +67,7 @@ public class ReviewDto {
         this.reviewId = reviewId;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
